Added a new Cog, for webscraping-related activities
This will be the main Cog for ArcaeaBot which manages and updates recent data on the GSheet instance. Now, I know using any sort of data sheet as a database is a bad idea, but I feel like having it being user-friendly(ish) to read is useful, so that people will be able to use the data regardless if they need the bot or not. Basically, this Bot's doing the job of two different python scripts, but in one command instead.

Cool, right?